Move ViewModels and ViewMo.Factories to ui package
