+++ app/src/main/java/com/shapeapp/shape/fragments/PublicFragment.kt
@@ -19,8 +19,8 @@ import com.shapeapp.shape.mockupsmakers.UsersMockups
 import com.shapeapp.shape.recyclerviewadapters.SmallCardRecyclerViewAdapter
 import com.shapeapp.shape.recyclerviewinterfaces.RecyclerViewCardClickListener
 import com.shapeapp.shape.repositories.CardsRepository
-import com.shapeapp.shape.viewmodels.PublicFragmentViewModel
-import com.shapeapp.shape.viewmodels.PublicFragmentViewModelFactory
+import com.shapeapp.shape.ui.PublicFragmentViewModel
+import com.shapeapp.shape.ui.PublicFragmentViewModelFactory
 import kotlinx.android.synthetic.main.fragment_public.*
 
 // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
